Split structure.py into two files
